services/account-service/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
PyJWT>=2.8
psycopg[binary]>=3.2
python-dotenv>=1.0
Pillow>=10.0
cloudinary>=1.36
django-prometheus>=2.3
